requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
requests>=2.31.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
selenium>=4.0.0
nltk>=3.8.1
webdriver_manager>=4.0.0
pyttsx3>=2.90
keyboard>=0.13.5
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python>=4.8.0
pywin32>=305